+++ cmd/cronosd/cmd/root.go
@@ -135,6 +135,7 @@ func initRootCmd(rootCmd *cobra.Command, encodingConfig params.EncodingConfig) {
 		queryCommand(),
 		txCommand(),
 		ethermintclient.KeyCommands(app.DefaultNodeHome),
+		FixUnluckyTxCmd(),
 	)
 
 	// add rosetta